init: split and rename module, message
